rename: Game315Work로 파일명 및 폴더명 변경, Game315 컴포넌트 폴더 위치 변경 (CC-105)
